+++ pulp/0.2/label_balance_edges.cpp
@@ -63,9 +63,9 @@ void label_balance_edges(pulp_graph_t& g, int64_t num_parts, int64_t* parts,
 
   int64_t* part_sizes = new int64_t[num_parts];
   unsigned* part_edge_sizes = new unsigned[num_parts]; 
-  for (int i = 0; i < num_parts; ++i)
+  for (int64_t i = 0; i < num_parts; ++i)
     part_sizes[i] = 0;
-  for (int i = 0; i < num_parts; ++i)
+  for (int64_t i = 0; i < num_parts; ++i)
     part_edge_sizes[i] = 0;
 
   double avg_size = num_verts / num_parts;
@@ -93,22 +93,22 @@ void label_balance_edges(pulp_graph_t& g, int64_t num_parts, int64_t* parts,
 
   int64_t* part_sizes_thread = new int64_t[num_parts];
   unsigned* part_edge_sizes_thread = new unsigned[num_parts];
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
     part_sizes_thread[i] = 0;
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
     part_edge_sizes_thread[i] = 0;
   
 #pragma omp for schedule(static) nowait
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
   {
     ++part_sizes_thread[parts[i]];
     part_edge_sizes_thread[parts[i]] += out_degree(g, i);
   }
 
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
 #pragma omp atomic
     part_sizes[i] += part_sizes_thread[i];
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
 #pragma omp atomic
     part_edge_sizes[i] += part_edge_sizes_thread[i];
 
@@ -120,14 +120,14 @@ void label_balance_edges(pulp_graph_t& g, int64_t num_parts, int64_t* parts,
   double* part_edge_weights = new double[num_parts];
 
 #pragma omp for schedule(static) nowait
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
     in_queue_next[i] = false;
 
 while(t < edge_outer_iter)
 {
 
 #pragma omp for schedule(static)
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
     queue[i] = i;
 
 #pragma omp single
@@ -135,7 +135,7 @@ while(t < edge_outer_iter)
   num_swapped_1 = 0;
   queue_size = num_verts;
   next_size = 0;
-  for (int p = 0; p < num_parts; ++p)
+  for (int64_t p = 0; p < num_parts; ++p)
   {
     if ((double)part_edge_sizes[p] / avg_edge_size > max_e)
       max_e = (double)part_edge_sizes[p] / avg_edge_size;
@@ -145,7 +145,7 @@ while(t < edge_outer_iter)
   int64_t num_iter = 0;
   while (num_iter < edge_balance_iter)
   {
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       part_weights[p] = vert_balance * avg_size / (double)part_sizes[p] - 1.0;   
       part_edge_weights[p] = max_e * avg_edge_size / (double)part_edge_sizes[p] - 1.0;
@@ -156,12 +156,12 @@ while(t < edge_outer_iter)
     } 
 
 #pragma omp for schedule(guided) reduction(+:num_swapped_1) nowait
-    for (int i = 0; i < queue_size; ++i)
+    for (int64_t i = 0; i < queue_size; ++i)
     {
       int64_t v = queue[i];
       in_queue[v] = false;
       int64_t part = parts[v];
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0.0;
 
       unsigned out_degree = out_degree(g, v);
@@ -175,7 +175,7 @@ while(t < edge_outer_iter)
       
       int64_t max_part = part;
       double max_val = 0.0;
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
       {
         if (part_weights[p] > 0.0 && part_edge_weights[p] > 0.0)
           part_counts[p] *= (part_weights[p]*part_edge_weights[p]*weight_exponent_e);
@@ -213,7 +213,7 @@ while(t < edge_outer_iter)
             thread_start = next_size += thread_queue_size;
             
             thread_start -= thread_queue_size;
-            for (int l = 0; l < thread_queue_size; ++l)
+            for (int64_t l = 0; l < thread_queue_size; ++l)
               queue_next[thread_start+l] = thread_queue[l];
             thread_queue_size = 0;
           }
@@ -231,7 +231,7 @@ while(t < edge_outer_iter)
               thread_start = next_size += thread_queue_size;
               
               thread_start -= thread_queue_size;
-              for (int l = 0; l < thread_queue_size; ++l)
+              for (int64_t l = 0; l < thread_queue_size; ++l)
                 queue_next[thread_start+l] = thread_queue[l];
               thread_queue_size = 0;
             }
@@ -260,7 +260,7 @@ while(t < edge_outer_iter)
     thread_start = next_size += thread_queue_size;
     
     thread_start -= thread_queue_size;
-    for (int l = 0; l < thread_queue_size; ++l)
+    for (int64_t l = 0; l < thread_queue_size; ++l)
       queue_next[thread_start+l] = thread_queue[l];
     thread_queue_size = 0;
     
@@ -282,7 +282,7 @@ while(t < edge_outer_iter)
     next_size = 0;
 
     max_e = 0.0;
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       if ((double)part_edge_sizes[p] / avg_edge_size > max_e)
         max_e = (double)part_edge_sizes[p] / avg_edge_size;
@@ -303,7 +303,7 @@ while(t < edge_outer_iter)
   } // end while
 
 #pragma omp for schedule(static)
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
     queue[i] = i;
 
 #pragma omp single
@@ -316,7 +316,7 @@ while(t < edge_outer_iter)
   num_iter = 0;
   while (num_iter < edge_refine_iter)
   {
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       part_weights[p] = vert_balance * avg_size / (double)part_sizes[p] - 1.0;   
       part_edge_weights[p] = max_e * avg_edge_size / (double)part_edge_sizes[p] - 1.0;
@@ -327,11 +327,11 @@ while(t < edge_outer_iter)
     }
 
 #pragma omp for schedule(guided) reduction(+:num_swapped_2) nowait  
-    for (int i = 0; i < queue_size; ++i)
+    for (int64_t i = 0; i < queue_size; ++i)
     {
       int64_t v = queue[i];
       in_queue[v] = false;
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0;
 
       int64_t part = parts[v];
@@ -346,7 +346,7 @@ while(t < edge_outer_iter)
 
       int64_t max_part = -1;
       int64_t max_count = -1;
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
         if (part_counts[p] > max_count)
         {
           max_count = part_counts[p];
@@ -382,7 +382,7 @@ while(t < edge_outer_iter)
               thread_start = next_size += thread_queue_size;
               
               thread_start -= thread_queue_size;
-              for (int l = 0; l < thread_queue_size; ++l)
+              for (int64_t l = 0; l < thread_queue_size; ++l)
                 queue_next[thread_start+l] = thread_queue[l];
               thread_queue_size = 0;
             }
@@ -400,7 +400,7 @@ while(t < edge_outer_iter)
                 thread_start = next_size += thread_queue_size;
                 
                 thread_start -= thread_queue_size;
-                for (int l = 0; l < thread_queue_size; ++l)
+                for (int64_t l = 0; l < thread_queue_size; ++l)
                   queue_next[thread_start+l] = thread_queue[l];
                 thread_queue_size = 0;
               }
@@ -414,7 +414,7 @@ while(t < edge_outer_iter)
     thread_start = next_size += thread_queue_size;
     
     thread_start -= thread_queue_size;
-    for (int l = 0; l < thread_queue_size; ++l)
+    for (int64_t l = 0; l < thread_queue_size; ++l)
       queue_next[thread_start+l] = thread_queue[l];
     thread_queue_size = 0;
 
@@ -438,7 +438,7 @@ while(t < edge_outer_iter)
     num_swapped_2 = 0;  
 
     max_e = 0.0;
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       if ((double)part_edge_sizes[p] / avg_edge_size > max_e)
         max_e = (double)part_edge_sizes[p] / avg_edge_size;
@@ -514,9 +514,9 @@ void label_balance_edges_weighted(
 
   int64_t* part_sizes = new int64_t[num_parts];
   unsigned* part_edge_sizes = new unsigned[num_parts]; 
-  for (int i = 0; i < num_parts; ++i)
+  for (int64_t i = 0; i < num_parts; ++i)
     part_sizes[i] = 0;
-  for (int i = 0; i < num_parts; ++i)
+  for (int64_t i = 0; i < num_parts; ++i)
     part_edge_sizes[i] = 0;
 
   double avg_size = (double)g.vertex_weights_sum / (double)num_parts;
@@ -544,13 +544,13 @@ void label_balance_edges_weighted(
 
   int64_t* part_sizes_thread = new int64_t[num_parts];
   unsigned* part_edge_sizes_thread = new unsigned[num_parts];
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
     part_sizes_thread[i] = 0;
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
     part_edge_sizes_thread[i] = 0;
   
 #pragma omp for schedule(static) nowait
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
   {
     if (has_vwgts)
       part_sizes_thread[parts[i]] += g.vertex_weights[i];
@@ -559,10 +559,10 @@ void label_balance_edges_weighted(
     part_edge_sizes_thread[parts[i]] += out_degree(g, i);
   }
 
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
 #pragma omp atomic
     part_sizes[i] += part_sizes_thread[i];
-  for (int i = 0; i < num_parts; ++i) 
+  for (int64_t i = 0; i < num_parts; ++i) 
 #pragma omp atomic
     part_edge_sizes[i] += part_edge_sizes_thread[i];
 
@@ -574,14 +574,14 @@ void label_balance_edges_weighted(
   double* part_edge_weights = new double[num_parts];
 
 #pragma omp for schedule(static) nowait
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
     in_queue_next[i] = false;
 
 while(t < edge_outer_iter)
 {
 
 #pragma omp for schedule(static)
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
     queue[i] = i;
 
 #pragma omp single
@@ -589,7 +589,7 @@ while(t < edge_outer_iter)
   num_swapped_1 = 0;
   queue_size = num_verts;
   next_size = 0;
-  for (int p = 0; p < num_parts; ++p)
+  for (int64_t p = 0; p < num_parts; ++p)
   {
     if ((double)part_edge_sizes[p] / avg_edge_size > max_e)
       max_e = (double)part_edge_sizes[p] / avg_edge_size;
@@ -599,7 +599,7 @@ while(t < edge_outer_iter)
   int64_t num_iter = 0;
   while (num_iter < edge_balance_iter)
   {
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       part_weights[p] = vert_balance * avg_size / (double)part_sizes[p] - 1.0;   
       part_edge_weights[p] = max_e * avg_edge_size / (double)part_edge_sizes[p] - 1.0;
@@ -610,7 +610,7 @@ while(t < edge_outer_iter)
     } 
 
 #pragma omp for schedule(guided) reduction(+:num_swapped_1) nowait
-    for (int i = 0; i < queue_size; ++i)
+    for (int64_t i = 0; i < queue_size; ++i)
     {
       int64_t v = queue[i];
       in_queue[v] = false;
@@ -618,7 +618,7 @@ while(t < edge_outer_iter)
       int64_t v_weight = 1;
       if (has_vwgts) v_weight = g.vertex_weights[v];
 
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0.0;
 
       unsigned out_degree = out_degree(g, v);
@@ -635,7 +635,7 @@ while(t < edge_outer_iter)
       
       int64_t max_part = part;
       double max_val = 0.0;
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
       {
         if (part_weights[p] > 0.0 && part_edge_weights[p] > 0.0)
           part_counts[p] *= (part_weights[p]*part_edge_weights[p]*weight_exponent_e);
@@ -673,7 +673,7 @@ while(t < edge_outer_iter)
             thread_start = next_size += thread_queue_size;
             
             thread_start -= thread_queue_size;
-            for (int l = 0; l < thread_queue_size; ++l)
+            for (int64_t l = 0; l < thread_queue_size; ++l)
               queue_next[thread_start+l] = thread_queue[l];
             thread_queue_size = 0;
           }
@@ -691,7 +691,7 @@ while(t < edge_outer_iter)
               thread_start = next_size += thread_queue_size;
               
               thread_start -= thread_queue_size;
-              for (int l = 0; l < thread_queue_size; ++l)
+              for (int64_t l = 0; l < thread_queue_size; ++l)
                 queue_next[thread_start+l] = thread_queue[l];
               thread_queue_size = 0;
             }
@@ -720,7 +720,7 @@ while(t < edge_outer_iter)
     thread_start = next_size += thread_queue_size;
     
     thread_start -= thread_queue_size;
-    for (int l = 0; l < thread_queue_size; ++l)
+    for (int64_t l = 0; l < thread_queue_size; ++l)
       queue_next[thread_start+l] = thread_queue[l];
     thread_queue_size = 0;
     
@@ -742,7 +742,7 @@ while(t < edge_outer_iter)
     next_size = 0;
 
     max_e = 0.0;
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       if ((double)part_edge_sizes[p] / avg_edge_size > max_e)
         max_e = (double)part_edge_sizes[p] / avg_edge_size;
@@ -763,7 +763,7 @@ while(t < edge_outer_iter)
   } // end while
 
 #pragma omp for schedule(static)
-  for (int i = 0; i < num_verts; ++i)
+  for (int64_t i = 0; i < num_verts; ++i)
     queue[i] = i;
 
 #pragma omp single
@@ -776,7 +776,7 @@ while(t < edge_outer_iter)
   num_iter = 0;
   while (num_iter < edge_refine_iter)
   {
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       part_weights[p] = vert_balance * avg_size / (double)part_sizes[p] - 1.0;   
       part_edge_weights[p] = max_e * avg_edge_size / (double)part_edge_sizes[p] - 1.0;
@@ -787,7 +787,7 @@ while(t < edge_outer_iter)
     }
 
 #pragma omp for schedule(guided) reduction(+:num_swapped_2) nowait  
-    for (int i = 0; i < queue_size; ++i)
+    for (int64_t i = 0; i < queue_size; ++i)
     {
       int64_t v = queue[i];
       in_queue[v] = false;
@@ -795,7 +795,7 @@ while(t < edge_outer_iter)
       int64_t v_weight = 1;
       if (has_vwgts) v_weight = g.vertex_weights[v];
 
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
         part_counts[p] = 0;
 
       unsigned out_degree = out_degree(g, v);
@@ -812,7 +812,7 @@ while(t < edge_outer_iter)
 
       int64_t max_part = -1;
       int64_t max_count = -1;
-      for (int p = 0; p < num_parts; ++p)
+      for (int64_t p = 0; p < num_parts; ++p)
         if (part_counts[p] > max_count)
         {
           max_count = part_counts[p];
@@ -848,7 +848,7 @@ while(t < edge_outer_iter)
               thread_start = next_size += thread_queue_size;
               
               thread_start -= thread_queue_size;
-              for (int l = 0; l < thread_queue_size; ++l)
+              for (int64_t l = 0; l < thread_queue_size; ++l)
                 queue_next[thread_start+l] = thread_queue[l];
               thread_queue_size = 0;
             }
@@ -866,7 +866,7 @@ while(t < edge_outer_iter)
                 thread_start = next_size += thread_queue_size;
                 
                 thread_start -= thread_queue_size;
-                for (int l = 0; l < thread_queue_size; ++l)
+                for (int64_t l = 0; l < thread_queue_size; ++l)
                   queue_next[thread_start+l] = thread_queue[l];
                 thread_queue_size = 0;
               }
@@ -880,7 +880,7 @@ while(t < edge_outer_iter)
     thread_start = next_size += thread_queue_size;
     
     thread_start -= thread_queue_size;
-    for (int l = 0; l < thread_queue_size; ++l)
+    for (int64_t l = 0; l < thread_queue_size; ++l)
       queue_next[thread_start+l] = thread_queue[l];
     thread_queue_size = 0;
 
@@ -904,7 +904,7 @@ while(t < edge_outer_iter)
     num_swapped_2 = 0;  
 
     max_e = 0.0;
-    for (int p = 0; p < num_parts; ++p)
+    for (int64_t p = 0; p < num_parts; ++p)
     {
       if ((double)part_edge_sizes[p] / avg_edge_size > max_e)
         max_e = (double)part_edge_sizes[p] / avg_edge_size;
